+++ src/gfx/OpenGL/obj_loader (Tim-PC's conflicted copy 2015-08-29).cpp
@@ -377,7 +377,7 @@ static inline unsigned int ParseOBJIndexValue(const std::string& token, unsigned
 
 static inline float ParseOBJFloatValue(const std::string& token, unsigned int start, unsigned int end)
 {
-    return (float)atof(token.substr(start, end - start).c_str());
+    return atof(token.substr(start, end - start).c_str());
 }
 
 static inline std::vector<std::string> SplitString(const std::string &s, char delim)
