+++ src/globals (Tim-PC's conflicted copy 2015-08-29).cpp
@@ -2,7 +2,7 @@
 
 Globals::Globals()
 {
-	srand((unsigned int)time(NULL));
+	srand(time(NULL));
 }
 
 Globals::~Globals()
